requirements.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
decorator==4.2.1
Django==1.10
django-simple-history==1.9.0
djangorestframework==3.7.7
enum34==1.1.6
ipdb==0.11
ipython==5.5.0
ipython-genutils==0.2.0
pathlib2==2.3.0
pexpect==4.4.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
scandir==1.7
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.2.3
traitlets==4.3.2
wcwidth==0.1.7
